[KAN-76] refactor: filtering 및 sorting 로직들 useGatherings 훅으로 분리
