+++ src/app/(main)/gatherings/_component/ClientSideGatherings.tsx
@@ -8,108 +8,26 @@ import Tabs from '@/app/components/Tabs/Tabs';
 import Chips from '@/app/components/Chips/Chips';
 import CreateGatheringButton from './CreateGatheringButton';
 import MakeGatheringModal from '@/app/components/Modal/MakeGatheringModal';
+import useGatherings from '@/hooks/useGatherings';
 import usePreventScroll from '@/hooks/usePreventScroll';
-import getGatherings from '@/app/actions/gatherings/getGatherings';
 import { GatheringsListData } from '@/types/data.type';
-import { formatingDate } from '@/utils/formatDate';
-import { sortOptionsMap } from '@/utils/sortOptions';
 
 interface ClientSideGatheringsProps {
   gatherings: GatheringsListData[];
 }
 
 const ClientSideGatherings = ({ gatherings }: ClientSideGatheringsProps) => {
-  const [activeTab, setActiveTab] = useState<'WORKATION' | 'DALLAEMFIT'>(
-    'DALLAEMFIT',
-  );
   const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
-  const [filteredData, setFilteredData] =
-    useState<GatheringsListData[]>(gatherings);
-  const [selectedLocation, setSelectedLocation] = useState<string | undefined>(
-    undefined,
-  );
-  const [selectedChip, setSelectedChip] = useState<
-    'ALL' | 'OFFICE_STRETCHING' | 'MINDFULNESS' | null
-  >(null);
-  const [selectedDate, setSelectedDate] = useState<Date | null>(null);
-  const [sortOption, setSortOption] = useState<string | undefined>();
 
-  const handleTabClick = async (type: 'WORKATION' | 'DALLAEMFIT') => {
-    setActiveTab(type);
-    setSelectedChip(null);
-
-    const newData = await getGatherings({
-      type,
-      location: selectedLocation,
-      date: selectedDate ? formatingDate(selectedDate) : undefined,
-      sortBy: sortOption,
-    });
-
-    setFilteredData(newData);
-  };
-
-  const handleChipClick = async (
-    label: 'ALL' | 'OFFICE_STRETCHING' | 'MINDFULNESS',
-  ) => {
-    setSelectedChip(label);
-
-    const type = label === 'ALL' ? 'DALLAEMFIT' : label;
-
-    const newData = await getGatherings({
-      type,
-      location: selectedLocation,
-      date: selectedDate ? formatingDate(selectedDate) : undefined,
-      sortBy: sortOption,
-    });
-
-    setFilteredData(newData || []);
-  };
-
-  const handleLocationChange = async (location: string | undefined) => {
-    setSelectedLocation(location);
-
-    const type =
-      selectedChip === 'ALL' || !selectedChip ? activeTab : selectedChip;
-    const newData = await getGatherings({
-      type,
-      location,
-      date: selectedDate ? formatingDate(selectedDate) : undefined,
-      sortBy: sortOption,
-    });
-
-    setFilteredData(newData || []);
-  };
-
-  const handleDateChange = async (date: Date | null) => {
-    setSelectedDate(date);
-
-    const type =
-      selectedChip === 'ALL' || !selectedChip ? activeTab : selectedChip;
-    const newData = await getGatherings({
-      type,
-      location: selectedLocation,
-      date: date ? formatingDate(date) : undefined,
-      sortBy: sortOption,
-    });
-
-    setFilteredData(newData || []);
-  };
-
-  const handleSortChange = async (option: string) => {
-    const sortBy = sortOptionsMap[option] || undefined;
-    setSortOption(sortBy);
-
-    const type =
-      selectedChip === 'ALL' || !selectedChip ? activeTab : selectedChip;
-    const newData = await getGatherings({
-      type,
-      location: selectedLocation,
-      date: selectedDate ? formatingDate(selectedDate) : undefined,
-      sortBy,
-    });
-
-    setFilteredData(newData || []);
-  };
+  const {
+    filteredData,
+    activeTab,
+    handleTabClick,
+    handleChipClick,
+    handleLocationChange,
+    handleDateChange,
+    handleSortChange,
+  } = useGatherings(gatherings);
 
   usePreventScroll(isModalOpen);
